pyproject.toml
[project]
name = "notifications-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.128.0",
    "loguru==0.7.3",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.12.5",
    "python-decouple==3.8",
    "python-jose[cryptography]==3.5.0",
    "redis==7.1.0",
    "tortoise-orm[asyncpg]==0.25.3",
]

[project.optional-dependencies]
test = [
    "httpx==0.28.1",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
]

